Refactor log header
